# TP7_API_LARAVEL_INFYOM/storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/clientphysiques": {
            "get": {
                "tags": [
                    "Clientphysique"
                ],
                "summary": "Get a listing of the Clientphysiques.",
                "description": "Get all Clientphysiques",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Clientphysique"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientphysique"
                ],
                "summary": "Store a newly created Clientphysique in storage",
                "description": "Store Clientphysique",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Clientphysique that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Clientphysique"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Clientphysique"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/clientphysiques/{id}": {
            "get": {
                "tags": [
                    "Clientphysique"
                ],
                "summary": "Display the specified Clientphysique",
                "description": "Get Clientphysique",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Clientphysique",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Clientphysique"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientphysique"
                ],
                "summary": "Update the specified Clientphysique in storage",
                "description": "Update Clientphysique",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Clientphysique",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Clientphysique that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Clientphysique"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Clientphysique"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientphysique"
                ],
                "summary": "Remove the specified Clientphysique from storage",
                "description": "Delete Clientphysique",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Clientphysique",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Clientphysique": {
            "required": [
                "numId",
                "nom",
                "prenom",
                "email",
                "cni",
                "telephone",
                "adresse",
                "sexe",
                "dateNaiss",
                "dateCreation",
                "features",
                "isSalarie"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "numId": {
                    "description": "numId",
                    "type": "string"
                },
                "nom": {
                    "description": "nom",
                    "type": "string"
                },
                "prenom": {
                    "description": "prenom",
                    "type": "string"
                },
                "email": {
                    "description": "email",
                    "type": "string"
                },
                "cni": {
                    "description": "cni",
                    "type": "string"
                },
                "telephone": {
                    "description": "telephone",
                    "type": "string"
                },
                "adresse": {
                    "description": "adresse",
                    "type": "string"
                },
                "sexe": {
                    "description": "sexe",
                    "type": "string"
                },
                "dateNaiss": {
                    "description": "dateNaiss",
                    "type": "string"
                },
                "dateCreation": {
                    "description": "dateCreation",
                    "type": "string"
                },
                "features": {
                    "description": "features",
                    "type": "string"
                },
                "isSalarie": {
                    "description": "isSalarie",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                },
                "deleted_at": {
                    "description": "deleted_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    }
}